+++ src/pyvider/components/capabilities/lens.py
@@ -9,7 +9,6 @@
 from typing import Any
 
 from provide.foundation import logger
-
 from pyvider.capabilities import BaseCapability, register_capability
 from pyvider.cty import CtyValue
 from pyvider.cty.conversion import infer_cty_type_from_raw